fix commands
