+++ packages/@dcl/sdk/package.json
@@ -34,7 +34,9 @@
     "@well-known-components/logger": "^3.0.0",
     "@well-known-components/metrics": "^2.0.1-20220909150423.commit-8f7e5bc",
     "arg": "5.0.2",
+    "colorette": "^2.0.19",
     "extract-zip": "2.0.1",
+    "ignore": "^5.2.4",
     "inquirer": "^8.2.5",
     "node-fetch": "^2.6.8",
     "open": "^8.4.0",
